Feature 4 / Menu system
Updates to SQL, added a version without username & password, just incase

A simple menu showing the 6 options now displays when ran.

Starting Feature 4, method displayaddEmployee opens a new menu letting the user add an employee to the DB.

As of now all String inputs are taken from 1 validation method. This method passes a requestMessage holding the question asked of the user.

 If the input passed the string regex, containing only letters, then the validated string is returned.

Date of birth also taken from a validation method, this tries to take the user input and parse it with the format 'yyyy-MM-dd' and catches the expection.

Next Commit will have:
1. All inouts validated
2. Employee data posted to database.
3. Implemented into the main menu.